Added relevant tables and graphs
Implemented and added relevant tables and graphs to the project. Included a bar graph illustrating the number of crimes committed in Toronto neighbourhoods, as well as scatter plots examining the correlations between the number of crimes and both population and income in different neighbourhoods. Applied appropriate styling, including plot titles, axis labels, and data labels. All visualizations are now part of the project to enhance data exploration and analysis.